backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
alembic==1.13.1
textblob==0.17.1
vaderSentiment==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-reddit==1.0.1
tweepy==4.14.0
fredapi==0.5.1
yfinance==0.2.36
celery==5.3.6
redis==5.0.1
pytest==8.0.0
httpx==0.26.0 